option zoom.left to allow zoom on the right side
Option to zoom to right seems to be a frequent request. This PR is partly based on https://stackoverflow.com/questions/52665619/how-to-change-the-position-of-the-zoomed-area-from-facet-zoom but added as an option. 
Not a fancy way to do it but just a number of ifelse statements. worked in my test environment (windows)
Would be great if I could add a wee bit to you great work, making our lives so much easier!